Apply sqlite to API, add /stake endpoints, remove stake estimates from /block endpoints.
Remove stake estimates and next diff from data storage as it is not very useful to lookup these values for past blocks, and we can't easily recompute them.

Make appContext.BlockData into an interface, APIDataSource, with the necessary methods for the API.

Add apisource.go to dcrsqlite package with wiredDB type to implement APIDataSource with the help of an RPC client for the block headers, which we don't store (in sql at least).

Add /stake, /stake/diff and /stake/estimates API endpoints.

Add dbfile config option for dcrdata.

Use logger for dcrslite in dcrdata and rebuilddb.

Fix dumb bugs in Store* methods of dcrsqlite.DB.

Add best summary height and stakeinfo height to dcrsqlite.DB as a cache so it's not necessary to query for it.

Add utilities to rpcutils:  GetStakeDiffEstimates, GetBlockHeaderVerbose, BuildBlockHeaderVerbose.