Make a collectionQueue to run certain collection/storage functions in order and synchronously.
This still allows the OnBlockConnected handler to return quickly and not block the RPC client.

Create the collectionQueue in make/getNodeNtfnHandlers and return it to main so that the synchronous collection functions can be set once they are available.

No loner send to the block connected channels for  stakedb, wiredDB, or blockdata, instead calling the new synchronous functions  BlockConnectedSync of each package's chain monitor.  The individual block connected channels are still used in BlockConnectedSync, but signals are used to wait until handling of the signaled block has completed in the old hander loops.

Create blockdata.(*Collector).CollectHash(*chainhash.Hash) to be used in the event that the chain node best block is higher than the block requested for data collection.  Pool value may be wrong if the stake DB is not on the same height however.  But the new synchronous execution changes are indended to prevent the stake DB from getting ahead.

Update blockdata.(*chainMonitor).BlockConnectedHandler so that it will call Collect() only if chain height equals the height of the requested block data, and call CollectHash(hash) if collection is behind.  A caveat with CollectHash, in addition to the pool value bit described above, is that it will not get the next block window estimates, however this is not needed if a higher block is waiting to be processed subsequently.

Change reorg handling in dcrsqlite so that instead of making a slice of block hashes for the new main chain and then trying to collect data for each, which is bad because the stake DB will be at tip and pool value for intermediate blocks will not be available, instead collect with CollectAPITypes and make a slice of  the the data.  The new blockAPISourceData type holds this data.  Then in switchToSideChain, just unpack the data.