collectionQueue to run certain collection functions synchronously and in order (#84)
* Make a collectionQueue to run certain collection/storage functions in order and synchronously.

This still allows the OnBlockConnected handler to return quickly and not block the RPC client.

Create the collectionQueue in make/getNodeNtfnHandlers and return it to main so that the synchronous collection functions can be set once they are available.

No loner send to the block connected channels for  stakedb, wiredDB, or blockdata, instead calling the new synchronous functions  BlockConnectedSync of each package's chain monitor.  The individual block connected channels are still used in BlockConnectedSync, but signals are used to wait until handling of the signaled block has completed in the old hander loops.

Create blockdata.(*Collector).CollectHash(*chainhash.Hash) to be used in the event that the chain node best block is higher than the block requested for data collection.  Pool value may be wrong if the stake DB is not on the same height however.  But the new synchronous execution changes are indended to prevent the stake DB from getting ahead.

Update blockdata.(*chainMonitor).BlockConnectedHandler so that it will call Collect() only if chain height equals the height of the requested block data, and call CollectHash(hash) if collection is behind.  A caveat with CollectHash, in addition to the pool value bit described above, is that it will not get the next block window estimates, however this is not needed if a higher block is waiting to be processed subsequently.

* dcrsqlite.DB locking for dbSummaryHeight and dbStakeInfoHeight

* fix deadlock.  need to rethink sqlite locking

* Create PoolInfoCache, and collect pool info for each block that is connected in stakedb.

* log when block not found in pool info cache

* More stern warning about pool info cache issues during sync

* Return to storing only block hash when building side chain slice.

stakedb does not connect side chain blocks until entire side chain is ready, which requires dcrsqlite to wait for the new best block to be processed by stakedb's handlers before attempting to get pool value for a given block on the side chain (new main chain)

This only works out because of the PoolInfoCache where stakedb will store pool info for connected blocks.

* document pool info get and cache set after stakedb block connect.

* rename daemonLog -> blockdataLog, DCRD -> BLKD

* Enable blockdata's reorg handler to collect update web UI only on last block of new main chain.

Add a special BlockDataSaver slice just for use in a reorg, so as not to run the sqlite saver twice.  dcrsqlite handles the reorg itself.  Turn trace logging back to info in blockdata.

Signal mempool collector that a new block was mined via a nil Hash, sent from (*collectionQueue).ProcessBlocks().  Update (*mempoolMonitor).TxHandler to recognize the signal and skip transaction checking and other irrelevant logic.  Refactor collection and storage code into (*mempoolMonitor) CollectAndStore() for reuse.

* doc NewCollectionQueue

closes Issue 97 (https://github.com/dcrdata/dcrdata/issues/97)