Add a block data cache, APICache. (#176)
* apicache: Add a block data cache, APICache.

APICache implements BlockSummarySaver, which takes a *BlockDataBasic
and caches it. A block priority queue, BlockPriorityQueue, is maintained
by APICache to decide which blocks to evict if needed. The "LessFn" can
be set to an arbitrary comparator of two CachedBlock values. There are
methods to remove and retrieve blocks from the cache.

* Do not error if MainchainBlocks too small, just grow.

Add docs on how to make the APICache and initialize MainchainBlocks.

* Define BlockSummarySaver