Switch from glide to dep (#86)
* dep init

* Add goclean.sh and address linters.

Enable goclean.sh in travis

install gometalinter in Travis

* update all deps.  Need to test chi, CORS, and sqlite

Add go-humanize to Gopkg.lock

* advance go-sqlite3 to master and fix it there

* address linters, clean up

The BlockPriorityQueue mutex was being copied because of sort.Interface using value receivers.  Make the mutex a pointer, and create it in BlockPriorityQueue contructor.

Reorder fields in BlockPriorityQueue to minimize padding.

BlockHashPathOrIndexCtx had an unchecked error.  This should be tested.

* Update dcrd and dcrwallet to v1.1.0.  Vote version and agendas are now updated.

* Remove the "/*" from each route in /api/list.

This was added to chi's mux.Routes() return sometime between v3.0.0 and v3.2.1.  Not sure why.