Query and cache for total address row count.
The address page uses a LIMIT query behind the scenes, so for the address page to say how many are in total, a second query is used to count the rows.  This is only necessary occasionally, so this adds the addressCounter type to store these counts.

New blocks invalidate (clear) the address counts and the queries must be executed again for fresh values.

explorer.AddressHistory now also returns the total number of funding/receiving transactions (total rows from the PSQL table).
Add KnownFundingTxns to explorer.AddressInfo and store the count from AddressHistory in it.

The dcrsqlite db (wiredDB) cannot get this count, so it is always 0 in the address page data structure when in lite mode.

Add NumFundingTxns and NumSpendingTxns to the explorer.AddressInfo type and set these values in both sqlite and pg dbs.
Remove AddressRow field.