tx page: Always show any decoded addresses, toggle OP_RETURN display.
If an output has an address, always display it, and if there is also OP_RETURN data, show a "toggle OP_RETURN" txt to display the data on click.
If an output has no decoded address, but OP_RETURN, show it like before.
Unspendable 0 value outputs marked as n/a in Spent column.