http profiling: fix mount of default http mux as a subroute (#359)
This adds a new config option --httpprofprefix to specify the URL path prefix for the HTTP profiler. For example, given -p --httpprofprefix "/pp", the http hander will be available at:

/pp/debug/pprof/
Note that the "/debug/pprof/ part is not changeable, and the trailing slash is required. But this path may be hidden via the prefix.