config: add appdata/datadir options (#371)
Add appdata and datadir options to config, and reorganize database folders. The config file may be:

* missing
* default location (in appdata)
* specified on command line with -C

If the config file is set on the command line, and it is not present, then it is an error.
Log both the log directory and the config file path.

Templates folder (views folder with tmpl files) must be in same folder as the executable.

appdata folder layout looks like this now:

.
├── data
│   └── mainnet
│       ├── dcrdata.sqlt.db
│       ├── stakenodes
│       │   └── metadata
│       │       ├── 000006.log
│       │       ├── 000008.ldb
│       │       ├── 000009.ldb
│       │       ├── 000010.ldb
│       │       ├── 000011.ldb
│       │       ├── CURRENT
│       │       ├── LOCK
│       │       ├── LOG
│       │       └── MANIFEST-000007
│       ├── ticket_pool.db
│       └── ticket_pool.db.lock
├── dcrdata.conf
└── logs
    └── mainnet
        └── dcrdata.log

Also adds nil pointer check on explorer.New.  This will happen if the any of the html template files are missing.