explorer: add websocket tx send buffer
This add a buffer ,WebsocketHub.newTxBuffer, of size newTxBufferSize
for the websocket new tx sends. The websocket clients are now of type
client ([]*MempoolTx and sync.RwMutex) to accommodate the bulk sends.
A ticker function, WebsocketHub.periodicBufferSend(), is sent on a
goroutine at the start of WebsocketHub.run() that sends the buffer every
bufferTickerInterval seconds. The buffer is cleared on every send.