explorer/websocket: avoid repeatedly unregistering a websocket client
The send and receive loops in RootWebsocket were both running
(*WebsocketHub).UnregisterClient.  In addition, RootWebsocket had a
defer running UnregisterClient. Only the defer is needed.  Remove the
other two.

In the receive loop, a warning was always issued when JSON.Receive
returned a non-nil error.  However, when a client is lost, "EOF" is
expected, so do not log in this case.