explorer: refactor the explorer templates so they are easier to manage (#395)
This simplifies the way the template files are handled by explorerUI by moving all the current template logic and types to templates.go where the templates are handled by the templates and pageTemplate types. Now adding a new template should be as simple as adding an name of the file sans "tmpl" extension into the tmpl slice in explorer.New.