Reorganize AddressInfo struct, document, and add KnownSpendingTxns
Add to AddressTx the field InOutID for the input or output index of the
relevant transaction input/output.

Add (a *AddressTx).IOID() to format a string for the tx input/output.

fix address credit query

Update TxnCount to return KnownTransactions, KnownFundingTxns, or
KnownSpendingTxns depending on TxnType.

Minimize use of .FullMode in address.tmpl

set ConfirmHeight in lite mode for address page