Consolidated debit view (#524)
This introduces a new "merged debit" view on the addresses history page that lists only unique transactions.  The Debit DCR amount is the sum of all outpoints corresponding to the given address that are spent in a transaction.

It also fixes a bug with the next button on the same page.