address page: Add charts (#597)
This bumps the dcrpg tables version to 3.5.0 from 3.4.1 with the
addition of the tx_type column to the addresses table, and the vins
table.  The upgrade is automatic and takes about 1 hr depending on
hardware.

This adds charts on the address page to display the following data
over time for the given address, and binned by several selectable
time frames:
- transaction type
- sent and received amounts
- unspent amount (a.k.a. balance)

The graphs use the standard Decred color scheme.

API endpoints are added for types and amount distribution charts.
The required data is retrieved by AJAX requests.