add the ability to use environment variables for configuration (#629)
main: Add the ability to use environment variables to set configuration.

The ability to set app configuration via environment variables makes this
more compatible with 12 factor app design, in addition to docker and 
docker-compose.

A simple config test file is also added.

A new import, the env package, is added that allows declaration of 
environment variables via struct tag. The go module files, and dep files
are updated to include the dependency.