dcrpg: add side chain queries
Add BlockStatus struct to describe the status of a block in the block
chain in terms of main/side, validity, height, and hashes of self and
next and previous blocks.
Add RetrieveSideChainBlocks to retrieve a slice of []BlockStatus for all
side chain blocks.

While here, stop shaddowing the named return err everywhere in
queries.go.  Also remove redundant rows.Close().