dcrpg/views: side chain blocks page
Mounted on /side is a list of side chain blocks with links to them,
their parents, and their children, if any.
Update /block page to indicate if the block is a side chain block, and
if it is approved by stakeholders.
Add MainChain to explorer.BlockBasic.
When looking up a block from hash while in full mode, use the ChainDB
query instead of an RPC so that reorganized/orphaned blocks may be
located.