dcrpg/views: add side chain queries and side chain blocks page (#654)
This adds side chain queries and a side chain blocks page.

Add BlockStatus struct to describe the status of a block in the block
chain in terms of main/side, validity, height, and hashes of self and
next and previous blocks.
Add RetrieveSideChainBlocks to retrieve a slice of []BlockStatus for all
side chain blocks.

While here, stop shadowing the named return err everywhere in
queries.go.  Also remove redundant rows.Close().

Mounted on /side is a list of side chain blocks with links to them,
their parents, and their children, if any.
Update /block page to indicate if the block is a side chain block, and
if it is approved by stakeholders.
Add MainChain to explorer.BlockBasic.
When looking up a block from hash while in full mode, use the ChainDB
query instead of an RPC so that reorganized/orphaned blocks may be
located.