Current Ticket Pool Visualization (#542)
This add current ticket pool visualization graphs to /ticketpool page.

This also adds handling for the new getticketpooldata websocket messages.
TicketGrouping type and an enumeration is used to improve readability.
Rename 1d/1wk/1m to day/wk/mo. The "1" looks too much like an "l".
Use chaincfg.TicketMaturity for computing ticket maturity block.

The dcrpg table version is bumped to 3.5.2 with the addition of an index on
pool_status in the tickets table.