views: style invalidated transactions on block page. (#663)
This styles invalidated transactions on the block and transactions pages.

Create an invalidated-tx CSS style to gray out and change background color
of regular transaction table data, including the coinbase, on a block's page.
On the blocks list page, indicate invalidated blocks.  When not in lite mode,
get the Valid and MainChain status from PG.
Also change maxExplorerRows to 400 (from 2000).