dcrpg: transaction queries prefer valid and mainchain
Order by is_valid and is_mainchain so that the first row will be a valid
and/or mainchain block transaction if there is one.
Remove unused functions from queries.go.
NOTE that some of these queries may need to return is_valid and
is_mainchain so that they can report it.