views/css: render the list of blocks for a transaction
When a transaction is found in multiple blocks, list them all with
unambiguous links by hash.  Also pass the transaction index via ?ind=
URL parameter so that we can use it on the block page to highlight
the transaction (not done in this work).
Create "grayed out" colors with a "grayed" class for <body> and <a>
elements.  Also define these for dark/night mode.
Also simplify the signature for several template functions.