Transaction block list (#666)
This shows on the transaction view a list of blocks containing that transaction.

There should be only one block that is both mainchain and stake-validated. Any other
blocks will be "Mainchain: false" or "Valid: false" or both.

When a transaction is found in multiple blocks, list them all with unambiguous links by
hash.  Also pass the transaction index via ?ind= URL parameter so that we can use it
on the block page to highlight the transaction (not done in this work). Create
"grayed out" colors with a "grayed" class for <body> and <a> elements.  Also define
these for dark/night mode. Also simplify the signature for several template functions.

To get the block list, this also adds a new query to get details for all blocks containing
a specified transaction. The details are block index, height, hash, is_valid, and
is_mainchain.

Fix some other queries that are intended to return just one row, but may return many.
Order by is_valid and is_mainchain so that the first row will be a valid and/or mainchain
block transaction if there is one.  Remove unused functions from queries.go.  NOTE
that some of these queries may need to return is_valid and is_mainchain so that they
can report it.