dcrpg/dcrsqlite: no transactions found for address is not an error
In both dcrpg (for full mode) and dcrsqlite (for lite mode), this
responds to address history requests without error when there are no
known transactions for a given address but no other error.

For the API, this means returning a non-nil *apitypes.Address with the
Transactions field set to a non-nil zero-length slice.  The slice must
be non-nil so that when marshalled to JSON, "address_transactions"
shows as [] instead of null.