dcrpg: ticket pool cache at dcrpg level
So the APIs and the explorer can easily leverage the cache but not
even know about it, the cache should live entirely within ChainDB in the
dcrpg package.  Also, consumers only need use TicketPoolVisualization,
not the cache query and update functions.
ticketPoolDataCache lives in pgblockchain.go (dcrpg) now, and it is
managed entirely by ChainDB.
Modify tickets pool by date api response to include height of the data.
Also pass ChartsHeight to the /ticketpool page template, for future use.
The ticketPoolVisualization function now ensures that the DB height
hasn't changed from the beginning of the queries to the end.
Move ChartDataCounter out of explorertypes.go, which is intended to help
importers of the explore package find useful exported types.

js: update ticketpool data format in ws and ajax handlers

dcrpg: separate db guards for each ticket pool interval

Each interval has its own current height. Never clear the charts data
for all chart intervals from the cache, only update.
Separate notions of stale and found. On startup, the cache will be
empty, so this is not found and callers must wait. When found, but
stale, the stale data may be returned.
Add thorough docs to TicketPoolVisualization.