explorer: lookup txn in db that is not known to dcrd
If a transaction is not known to dcrd, it is because it has been
orphaned by a reorg, invalidated via stakeholder disapproval, or both.
This change will attempt to load details of the transaction from the
dcrpg database, which may have decoded and recorded the transaction.
Add a message to the transaction page when the transaction is not found
in any valid and mainchain block.