dcrpg: update addresses table during stake invalidation
The new block handler neglected to update valid_mainchain=false on the
previous block.  This change adds UpdateLastAddressesValid, which sets
valid_mainchain for the regular transactions in a given block.  Call it
from the "if !lastIsValid" block of (*ChainDB).StoreBlock, where the
previous vins and transactoins were correctly being invalidated.
Modify RetrieveTxnsVinsVoutsByBlock so the caller may specify if only
regular transactions' vins/vouts are needed, or all transaction types.