dcrpg: combine spent and unspent totals queries into one
Add SelectAddressSpentUnspentCountAndValue query that gets the count
and total value for spent and unspent address table rows for a given
address.  It uses "GROUP BY (is_funding, matching_tx_hash='')" with
aggregate functions to do it efficiently in a single scan on the
addresses table.
Scanning the result, isFunding && noMatchingTx means unspent,
!isFunding && !noMatchingTx means spent, while !isFunding &&
noMatchingTx is not expected so an error is logged.

Also clean-up some query formatting.
Fix the conditional on the (unused) InsertAddressRowReturnID query.