semver/rpcutils: allow multiple compatible dcrd RPC API versions (#687)
This change to dcrdata introduces a slice of compatible dcrd RPC API
versions, instead of just one, and a new function to check that the
connected node API version is compatible with at least one of the
versions listed in the slice.

dcrd RPC API versions 3.0.0 and 4.0.0 are now compatible as defined
by semver.Compatible.

semver: Add semver.AnyCompatible to check if a Semver is compatible
with any in a []Semver.  Add tests for package semver.

rpcutils: Instead of a single dcrd RPC API version, there is now a slice of
semver.Semver values.  The semver.AnyCompatible function checks the
actual dcrd RPC API version against the slice of compatible versions.