Invalidated transaction display (#676)
If a transaction is not known to dcrd, it is because it has been orphaned by
a reorg, invalidated via stakeholder disapproval, or both. This change will
attempt to load details of the transaction from the dcrpg database, which
may have decoded and recorded the transaction.

This adds a message to the transaction page when the transaction is not
found in any valid and mainchain block.

Main changes:

- Load transaction details from DB rather than dcrd.
- Handle globally-invalidated/sidechain transactions with a message on the
  transaction page.
- Pull the vins and vouts from the db too.

Future: See if it is easier, more efficient, or feasible to pull the entire block from
dcrd instead of checking the database.

Code changed details:

* dcrpg: add queries for all transactions table rows by hash
* explorer: lookup txn in db that is not known to dcrd
* dcrpg: get vins and vouts for a tx from DB

Add VinsForTx and VoutsForTx to get the vin/vout dbtypes for a given
dbtypes.Tx.
Add RetrievePkScriptByOutpoint to get an outpoints pkScript from the
vouts table. Add query SelectPkScriptByOutpoint.
Fix pgblockchain tests.

* explorer: get vins/vouts from dcrpg too

Use VinsForTx and VoutsForTx, and convert from dbtypes to explorer type.

* txhelpers:

Add CoinbaseScript to txhelpers.