dcrpg: addresses upsert must update in/out data
Each of matching_tx_hash, matching_tx_index, tx_hash, and
tx_vin_vout_index need to be updated when performing an upsert in the
addresses table.  When a block is invalidated, the transactions are able
to be mined in a different block, in which case it is possible that the
matching_tx_hash and matching_tx_index entries may be change
for funding addresses rows.

Add query UpdateAddressesFundingMatchingHash to repair
matching_tx_hash using info from vins.is_valid rows.  TODO: The upgrade
and dcrpg scheme version bump.