dcrpg: patch addresses.matching_tx_hash/index
Due to the incorrect upsert statements, matching_tx_hash and
matching_tx_index in the addresses table may be incorrect if an
invalidate transaction was later re-mined and then validated.
This adds and upgrade (3.5.4 -> 3.5.5) that patches the matching_tx_hash
and matching_tx_index columns via the vins table.
This also avoids early return from an upgrade when the upgrade query
returns sql.ErrNoRows, as this is not necessarily an error.