dcrpg: error logging for (*ChainDB).SyncChainDB
If SyncChainDB fails to get a block from dcrd, the sync will hang
waiting to send the result back ot main, while dcrsqlite waits for the
block that it can't get.  This adds error logging in this case so it
is clear why caused dcrpg to end sync early, and that it is OK to
interrupt.