dcrpg: document and reorganize query functions (#698)
This documents most of the query functions in queries.go, and structures the file to
keep functions logically organized by table. The exceptions are some maintenance
functions that are moved to the top, and most of the Update functions that are
grouped at the bottom since updates generally happen in response to the same
events (stake invalidation or reorganization).

This also renames RetrieveAddressTxnOutputWithTransaction to
RetrieveAddressUTXOs, and RetrieveAllRevokesDbIDHashHeight to
RetrieveAllRevokes for clarity and readability. The variable totalMergedSpent had a
misleading name since it was a count, not an amount, so that was renamed to
numMergedSpent, and the amount/value variables were renamed from total* to amt*.