update addresses entries on stake-invalidation (#701)
dcrpg: patch matching_tx_hash, and update it during stake invalidation.

NOTE: This bumps dcrpg tables from 3.5.3 to 3.5.4, and a short (1-3 minutes)
automatic upgrade is required.

The addresses table upsert must update in/out data. Each of matching_tx_hash,
tx_hash, and tx_vin_vout_index need to be updated when performing an upsert
in the addresses table. When a block is invalidated, the transactions are able to
be mined in a different block, in which case it is possible that the
matching_tx_hash entry may be change for funding addresses rows.

Due to the incorrect upsert statements, matching_tx_hash in the addresses
table may be incorrect if an invalidate transaction was later re-mined and then
validated. This adds and upgrade (3.5.4 -> 3.5.5) that patches the
matching_tx_hash column via the vins table. This also avoids early return from
an upgrade when the upgrade query returns sql.ErrNoRows, as this is not
necessarily an error.