dcrpg: addresses table query optimization with block_time_index sorting (#706)
This creates the addresses' table block_time_index with sorting on the
block_time column in order to avoid a sort when querying for rows by a
particular address.
Add ReindexAddressesBlockTime, to simplify dropping and re-creating the
block_time_index with this sorting. Implement the upgrade from dcrpg
3.5.4 to 3.5.5 for this block_time reindexing.

This is motivated in part by the slowness of such queries with large numbers
of rows in the addresses table, such as
DshMNsvETDWpVoCe1re9NTAChiJagzsFV7J, which has over 116,746
transactions some of which have hundreds of inputs.
@papacarp Proposed the index sorting solution implemented here.

The reindexing is fairly quick, just a few minutes.