rpcutils: initial ChainTips helper with tests (#707)
This adds SideChains, which requests all branch tips in 
the block tree from dcrd, and returns only those tips 
that are side chains with valid headers or fully 
validated blocks. This corresponds to those elements in 
the response from the getchaintips RPC with status set 
to "valid-headers" or "valid-fork".

This will be used to identify and import data from side 
chains that may or may not have been part of a 
reorganization to which dcrdata was witness. See issue 
#338. The next steps for dcrdata are to get for each 
side chain the common ancestor with the main chain, and 
then add blocks to dcrpg as is_sidechain=true and 
is_valid={actual}.

A dcrdata/rpcclient update gets pulled in by updating the
required minimum rpcclient module version to v1.0.2 in
go.mod.

The test in rpcclient_online_test.go is more of an 
example that requires an actual dcrd connection. It is 
run with go test -tags testnetRPC.

The test in rpcclient_test.go checks the sideChainTips 
function, which filters a []dcrjson.GetChainTipsResult 
to a new []dcrjson.GetChainTipsResult that has only 
tips with status "valid-headers" or "valid-fork".