fix stakedb rewind error on startup
Fix 1: Rewinding of stakedb never goes below 0, do not use lastBlockPG,
which will be -1 when auxDB is empty. This addresses:
"failed to rewind stakedb: got 0, expecting -1"

Fix 2: Prevent the startHeight of the block scan in (*wiredDB).resyncDB
from beginning beyond fetchToHeight, as this is the height at which the
MasterBlockGetter in (*ChainDB).SyncChainDB begins notifying of new
blocks it fetches via RPC, and begins waiting for blocks to be connected
in stakedb in (*wiredDB).resyncDB.

Note that the stakedb rewind in main ensures stakedb is not beyond the
height of auxDB (a ChainDB).

Thoroughly document semantics of fetchToHeight and the relationship of
BlockGetter in baseDB and MasterBlockGetter in auxDB.

Add an early loop continue in (*wiredDB).resyncDB when that iteration's
only purpose was to advance the stakedb (catching up).