Load webserver soonest possible (#710)
This adds a "sync status" page while dcrdata is syncing its DBs with dcrd.

The sync status page shows progress bars for the main block data processing and the follow-up operations including indexing and updating spending information in the addresses table.

Updates to the syncing progress are pushed to the status page via a WebSocket.

The full explorer pages, not just the sync status page, may be loaded during sync depending on the new "sync-status-limit" switch.  This setting specifies a threshold on the number of blocks remaining to sync from dcrd, below which the explorer pages will be loadable instead of the sync status page.  By default, only the sync status page may load.

While the sync status page is up, the APIs return a JSON response indicating the status of the sync operation.

This also adds a desktop notification for when the sync is complete.