fix stakedb rewind error on startup (#720)
This fixes a bug when stakedb rewinds to 0, and documents full-mode sync architecture.

Rewinding of stakedb never goes below 0, so do not use lastBlockPG,
which will be -1 when auxDB is empty. This addresses the error:
"failed to rewind stakedb: got 0, expecting -1"

Prevent the startHeight of the block scan in (*wiredDB).resyncDB
from beginning beyond fetchToHeight, as this is the height at which the
MasterBlockGetter in (*ChainDB).SyncChainDB begins notifying of new
blocks it fetches via RPC, and begins waiting for blocks to be connected
in stakedb in (*wiredDB).resyncDB.

Add an early loop continue in (*wiredDB).resyncDB when that iteration's
only purpose was to advance the stakedb (catching up).