Register for dcrd notifications before waiting on monitors to exit (#726)
This fixes a bug caused by registering for dcrd notifications too late.

The call to RegisterNodeNtfnHandlers must come before waiting on the
monitors' WaitGroup.  A recent commit moved RegisterNodeNtfnHandlers
to after the monitors returned, leaving a call to RegisterNodeNtfnHandlers
immediately followed by program termination.

This PR moves RegisterNodeNtfnHandlers before starting the monitors
(and thus before waiting on the monitors to return).  Note that it should be
fine to register after starting the monitors as long as it is before wg.Wait().