show MaximumBlockSizes parameter on the /parameters page (#715)
This includes the MaximumBlockSizes parameter on the /parameters page.

Presently this will show the first element of the MaximumBlockSizes slice in the
chain parameters.  However, since this slice may have more than one element
when there is an agenda item to change the max block size, this edit will use
the first block size.  As a result, the block size may be out of date when and if
a relevant agenda passes and activates, thus making the second block size
correct.

When the `getblockchaininfo` RPC is available, this issue can be resolved.