dcrsqlite: properly compute min blocks to check in sync
minBlocksToCheck was computed using the "start height" of the sync loop,
which is normally one past the best block in the DB.  But it did not
include a necessary +1.  This change instead compares node height to DB
height in deciding if the DB is already sync'd or if the chain server is
behind.  startHeight may still be dbHeight or dbHeight+1 depending on
how on the state of aux DB coordinattion, but the computation of
minBlocksToCheck is now correct in any case.