dcrpg: fix panic while importing side chain block with missed vote
When importing side chain blocks, we do not have the "validators" list,
which indicates which votes are expected in the block, so no
missed votes are identified.  However, InsertVotes was checking to
ensure that the number of misses + votes is 5.  Disable this check
when the validators slice is empty (importing side chain block).