reduce messages for expected events on side chain block import (#732)
Certain conditions are presently expected when importing a side chain
block. Namely, there will be no ticket pool info, and the parent of the
side chain root block will be main chain. Do not log these cases.

This also silences an expected condition in a wiredDB query:
With empty block tables, the best block height will be returned as -1.
In this case, return from GetBestBlockSummary early before trying
and failing to get the block summary for this non-existent block. This
change is meant to further clean up log messages when there is no
real error.