with appdata set, update default config file to avoid error
With a missing config file in the default appdata directory,
it should not be an error. This updates the default config
file location when a custom appdata is set so there is no
error if the config file is missing.  It is only an error if the
config file location is explicitly set and does not exist.