Do not update API status when DB is empty.
When first starting with an empty sqlite DB, do not send a
bogus height (4294967295 = uint32(-1)) on the api status
update channel.

Add the block height to the error mesage if no block is
returned by appContext.BlockData.GetBestBlockSummary
when performing this lookup in response to an API status
update signal received by appContext.StatusNtfnHandler.