config: adjust config/log/data paths if appdata is changed (#736)
This adjusts the config/log/data paths if appdata is set.

With appdata set to a non-default value via either a
command line switch or the environment variable, this
updates the default config file location to avoid an error.

With a missing config file in the default appdata directory,
it should not be an error. This updates the default config
file location when a custom appdata is set so there is no
error if the config file is missing.  It is only an error if the
config file location is explicitly set and does not exist.

This also makes a small optimization such that the API
status is not updated with an invalid height when DB is
empty. When first starting with an empty sqlite DB, do not
send a bogus height (4294967295 = uint32(-1)) on the
API status update channel.

This also adds the block height to the error message if no
block is returned by
appContext.BlockData.GetBestBlockSummary when
performing this lookup in response to an API status update
signal received by appContext.StatusNtfnHandler.

The tests in config_test.go are reworked. Details:
- Use os.Unsetenv instead of setting DCRDATA_CONFIG_FILE to "".
- Make an empty appdata folder in TestMain for certain tests.
- Add disableConfigFileEnv to disable the config file setting made in
  TestMain for test that require no config file override, such as the
  appdata tests.
- Error check all loadConfig calls.
- Add appdata test to verify that setting appdata with no explicit
  change to config file location will change the default config file
  location.
- Add negative config file test to ensure loadConfig errors for a
  non-existent non-default (explicitly set) config file location.
- Reduce string literal repetition.
- Clean up and document more.
- Remove TestDefaultConfig and TestDefaultConfigFile since these
  tested no code at all.