dcrpg: StoreBlock takes updateExistingRecords flag
updateExistingRecords indicates that existing DB records should not be
updated. Such records might be for mainchain and/or valid  blocks, so
these imported blocks should not data in rows that are conflicting as
per the different table constraints and unique indexes.

In dcrpg's switchToSideChain, StoreBlock is called with
updateExisting=TRUE since these blocks are being added to the main
chain and the primary purpose of StoreBlock here is to update the
records.

In the import-side-chains loop in main(), StoreBlock is called with
updateExistingRecords=FALSE since these are side chain blocks,
and unlike a reorg, there are no main chain blocks that will follow
the addition of the side chain blocks.  The side chain entries
should not update, only insert when there are not existing records
as defined by the tables' constraints.

In SyncChainDB (batch store used to catch up on mainchain),
updateExistingRecords=TRUE since these are mainchain blocks.

The updateExistingRecords parameter is added to: InsertTxns,
InsertTickets, InsertVotes, InsertAddressRows, and
SetSpendingForFundingOP (all called from storeTxns).
Also updated with updateExisting: SetSpendingForVinDbID,
SetSpendingForVinDbIDs, insertAddrSpendingTxUpdateMatchedFunding,
and SetSpendingByVinID.

Most MakeXInsertStatement functions are modified with a second
parameter, updateExistingRecords, to choose between an upsert
(when true) and an insert-or-do-nothing (when false):
tickets, votes, misses, addresses, and transactions.
Block inserts are either insert or upsert because block hash is unique
and side chain import happens after main chain sync, ensuring that
any blocks inserted during import have current and exclusive side chain
status.

Instead of modifying the agendas insert statement factory, just skip
the agendas part of the loop in InsertVotes when updateExistingRecords
is false (and checked is true). There is no need for a insert or do
nothing kind of statement since updateExistingRecords=false implies
this is not a main chain block. There are no valid/main flags in the
agendas table to update anyway.

Fix MissingSideChainBlocks to start at lowest block in side chain,
and count up to the tip. Document the expected order in
rpcutils.SideChainFull.

Fix InsertVotes to always close both prepared statements and rollback
on an error.

Do not use ON CONSTRAINT since and index is not a constraint,
although a unique index effectively imposes a constraint.

Reorganize and document most of the statements in dcrpg/internal.
Standardize the various insert statements naming and documentation
across tables.

Make the InsertXConflictDoNothing kind of statements non-locking of
relevant rows.  The insert and select have a race, but there there
should be no concerns about making this atomic instead.

rebuilddb2 needs updateExistingRecord passed to StoreBlock