minor sync fixes (#738)
This contains a few minor sync fixes and some clean up.

In (*wiredDB).resyncDB, the following heights are kept updated: stakeDBHeight,
summaryHeight, and stakeInfoHeight.
The ticket pool info height check is now done even when found in cache.
The updateStatusChan was being sent the same value repeatedly. Changed to i,
the loop iteration. Move the chain height check to the last operation in the loop.
After the sync loop, call DBHeights one more time and display final heights, for
debugging.

Increase ticket pool cache size from 20 to 200 blocks over blocks behind.
In the secondary calls to getSyncd, auxDBHeight was not being updated.
Update it so fetchToHeight = auxDBHeight + 1 is correct even if there is a third
call to getSyncd.

Remove indent from chain monitor setup block. Something strange was happening
with the monitors and I do not want issues with them or garbage collection.